Create ServiceDefinition @Bean in ServiceConfiguration
