+++ src/main/java/org/cloudfoundry/community/servicebroker/s3/config/ServiceConfiguration.java
@@ -1,0 +1,45 @@
+package org.cloudfoundry.community.servicebroker.s3.config;
+
+import org.cloudfoundry.community.servicebroker.model.Plan;
+import org.cloudfoundry.community.servicebroker.model.ServiceDefinition;
+import org.cloudfoundry.community.servicebroker.s3.plan.basic.BasicPlan;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import java.io.IOException;
+import java.util.*;
+
+/**
+ * Created by jcarter on 22/01/16.
+ */
+@Configuration
+public class ServiceConfiguration {
+
+    @Bean
+    public ServiceDefinition serviceDefinition() throws IOException {
+        return new ServiceDefinition("s3", "amazon-s3",
+                "Amazon S3 is storage for the Internet.", true, getPlans(), getTags(), getServiceDefinitionMetadata(),
+                Arrays.asList("syslog_drain"), null);
+    }
+
+    private List<String> getTags() {
+        return Arrays.asList("s3", "object-storage");
+    }
+
+    private Map<String, Object> getServiceDefinitionMetadata() {
+        Map<String, Object> sdMetadata = new HashMap<String, Object>();
+        sdMetadata.put("displayName", "Amazon S3");
+        sdMetadata.put("imageUrl", "http://a1.awsstatic.com/images/logos/aws_logo.png");
+        sdMetadata.put("longDescription", "Amazon S3 Service");
+        sdMetadata.put("providerDisplayName", "Amazon");
+        sdMetadata.put("documentationUrl", "http://aws.amazon.com/s3");
+        sdMetadata.put("supportUrl", "http://aws.amazon.com/s3");
+        return sdMetadata;
+    }
+
+    private List<Plan> getPlans() {
+        List<Plan> myPlans = new ArrayList<Plan>();
+        myPlans.add(BasicPlan.getPlan());
+        return myPlans;
+    }
+}
